requirements.txt
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
cssselect==1.2.0
cssutils==2.6.0
geomet==0.2.1.post1
idna==3.4
lxml==4.9.2
premailer==3.10.0
pycparser==2.21
python-dotenv==1.0.0
requests==2.30.0
six==1.16.0
urllib3==2.0.2
yagmail==0.15.293
